+++ src/onegov.election_day/onegov/election_day/tests/utils/test_pdf_generator.py
@@ -63,6 +63,7 @@ def test_generate_pdf_election_compound(session, election_day_app):
 
     compound = add_election_compound(session)
     compound.show_party_strengths = True
+    compound.show_mandate_allocation = True
     for locale in ('de_CH', 'fr_CH', 'it_CH', 'rm_CH'):
         generator.generate_pdf(compound, 'election.pdf', locale)
         with election_day_app.filestorage.open('election.pdf', 'rb') as f: