+++ src/onegov.election_day/onegov/election_day/tests/views/test_views_election_compound.py
@@ -162,6 +162,47 @@ def test_view_election_compound_party_strengths(election_day_app_gr):
     ])
 
 
+def test_view_election_compound_mandate_allocation(election_day_app_gr):
+    client = Client(election_day_app_gr)
+    client.get('/locale/de_CH').follow()
+
+    login(client)
+    create_election_compound(client)
+    upload_party_results(client, slug='elections/elections')
+
+    main = client.get('/elections/elections/mandate-allocation')
+    assert '<h3>Sitzzuteilung</h3>' in main
+
+    # Historical data
+    csv_parties = (
+        'year,name,id,total_votes,color,mandates,votes\r\n'
+        '2015,BDP,0,60000,#efb52c,1,10000\r\n'
+        '2015,CVP,1,60000,#ff6300,1,30000\r\n'
+        '2015,FDP,2,60000,#4068c8,0,20000\r\n'
+        '2011,BDP,0,40000,#efb52c,1,1000\r\n'
+        '2011,CVP,1,40000,#ff6300,1,15000\r\n'
+        '2011,FDP,2,40000,#4068c8,1,10000\r\n'
+    ).encode('utf-8')
+
+    upload = client.get('/elections/elections/upload-party-results')
+    upload.form['parties'] = Upload('parties.csv', csv_parties, 'text/plain')
+    upload = upload.form.submit()
+    assert "erfolgreich hochgeladen" in upload
+
+    results = client.get('/elections/elections/mandate-allocation').text
+    assert '2.5%' in results
+    assert '16.7%' in results
+    assert '14.2%' in results
+
+    assert '37.5%' in results
+    assert '50.0%' in results
+    assert '12.5%' in results
+
+    assert '25.0%' in results
+    assert '33.3%' in results
+    assert '8.3%' in results
+
+
 def test_view_election_compound_parties_panachage(election_day_app_gr):
     client = Client(election_day_app_gr)
     client.get('/locale/de_CH').follow()